pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.0",
    "black>=26.3.1",
    "fastapi>=0.135.3",
    "greenlet>=3.3.2",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
